+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.0] - 2021-07-13
+### Added
+- Add option IsOutputJson for write log json friendly elasticsearch
+### Changed
+- Change AddMakeSimpleLoging to AddMakeSimpleLogingRequest
+
 ## [0.1.5] - 2021-05-02
 ### Added
 - Add field Timestamp